:bulb: 상품 이미지 Controller 수정
## Description
1. ProductImageApiController
  - checkUser() 메서드를 통해 자신이 등록한 상품이 맞는지 확인하기 위해 MemberContext 타입 매개변수 추가

2. ProductImageService
  - 자신이 등록한 상품이 아닌 경우 AnotherMemberException 예외가 발생하도록 로직 수정

+++ src/main/java/com/market/carrot/product/service/ProductImageServiceImpl.java
@@ -1,7 +1,10 @@
 package com.market.carrot.product.service;
 
+import com.market.carrot.global.Exception.AnotherMemberException;
 import com.market.carrot.global.Exception.ExceptionMessage;
 import com.market.carrot.global.Exception.NotFoundEntityException;
+import com.market.carrot.login.config.customAuthentication.common.MemberContext;
+import com.market.carrot.product.domain.Product;
 import com.market.carrot.product.domain.ProductImage;
 import com.market.carrot.product.domain.ProductImageRepository;
 import com.market.carrot.product.dto.request.UpdateProductImageRequest;
@@ -18,18 +21,33 @@ public class ProductImageServiceImpl implements ProductImageService {
 
   @Transactional
   @Override
-  public void updateImage(Long id, UpdateProductImageRequest imageRequest) {
+  public void updateImage(Long id, UpdateProductImageRequest imageRequest,
+      MemberContext memberContext) {
     ProductImage findProductImage = productImageRepository.findById(id)
-        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_IMAGE, HttpStatus.BAD_REQUEST));
+        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_IMAGE,
+            HttpStatus.BAD_REQUEST));
+
+    Product findProductByImage = findProductImage.getProduct();
+
+    if (!findProductByImage.checkUser(memberContext.getMember())) {
+      throw new AnotherMemberException(ExceptionMessage.IS_NOT_WRITER, HttpStatus.BAD_REQUEST);
+    }
 
     findProductImage.updateProductImage(imageRequest);
   }
 
   @Transactional
   @Override
-  public void deleteImage(Long id) {
+  public void deleteImage(Long id, MemberContext memberContext) {
     ProductImage findProductImage = productImageRepository.findById(id)
-        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_IMAGE, HttpStatus.BAD_REQUEST));
+        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_IMAGE,
+            HttpStatus.BAD_REQUEST));
+
+    Product findProductByImage = findProductImage.getProduct();
+
+    if (!findProductByImage.checkUser(memberContext.getMember())) {
+      throw new AnotherMemberException(ExceptionMessage.IS_NOT_WRITER, HttpStatus.BAD_REQUEST);
+    }
 
     productImageRepository.delete(findProductImage);
   }
